pyproject.toml
[project]
name = "validator-template"
version = "0.0.0"
description = "Template repo for Guardrails Hub validators."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest>=7.4",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]
